scanner-worker/requirements.txt
boto3==1.34.25
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-decouple==3.8
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==4.2.14
